fix multiple concurrent tasks
